bana/requirements.txt
asgiref==3.8.1
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
configparser==5.0.2
crispy-bootstrap5==2023.10
cryptography==44.0.2
Django==4.2.20
django-allauth==65.7.0
django-browser-reload==1.12.1
django-ckeditor==6.7.1
django-classy-tags==4.1.0
django-cms==4.1.4
django-cors-headers==4.3.1
django-cors-middleware==1.5.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-formtools==2.5.1
django-htmx==1.17.2
django-js-asset==2.2.0
django-livesync==0.5
django-phonenumber-field==6.0.0
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-text-ckeditor==5.1.6
googlemaps==4.10.0
html5lib==1.1
idna==3.10
install==1.3.5
jsonfield==3.1.0
numpy==1.24.4
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.12.39
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
pytz==2021.3
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
urllib3==2.2.3
watchdog==4.0.0
webencodings==0.5.1
websocket-client==1.7.0
